Align public Scan selector and typing contracts
Require string selectors for Scan across the portable and qualified
surfaces, and normalize shared aliases through one path. Keep callable
operators as a qualified-only extension.

Use common read-only payload, storage, and contextual initial-value
typing contracts, and remove the public BlockScanAlgorithm enum.

+++ python/cuda_coop/tests/packaging/typing/portable_consumer.py
@@ -152,7 +152,7 @@ def check_portable_surface(source: object, destination: object) -> None:
             logical_warp,
             np.float32(4),
             scan_op="multiplies",
-            initial_value=np.float32(1),
+            initial_value=1.0,
         ),
         np.float32,
     )
